Rename transfrom/transform.js to transform/default.js
This js module contains the default transform which is a noop transform.
Therefore use a more self explaining name.
