+++ ng/src/gmp/http/transform/default.js
